add clabckas and make them fit
