+++ examples/mnist_fit.py
@@ -3,6 +3,7 @@ from typing import Any, Generator, Mapping, Tuple
 import haiku as hk
 import jax
 import jax.numpy as jnp
+import matplotlib.pyplot as plt
 import numpy as np
 import tensorflow_datasets as tfds
 import typer
@@ -45,6 +46,23 @@ def net_fn(image) -> jnp.ndarray:
     return mlp(image)
 
 
+def plot_history(history):
+    n_plots = len(history.history.keys()) // 2
+    plt.figure(figsize=(8, 12))
+    for i, key in enumerate(list(history.history.keys())[:n_plots]):
+        metric = history.history[key]
+        val_metric = history.history[f"val_{key}"]
+
+        plt.subplot(n_plots, 1, i + 1)
+        plt.plot(metric, label=f"Training {key}")
+        plt.plot(val_metric, label=f"Validation {key}")
+        plt.legend(loc="lower right")
+        plt.ylabel(key)
+        plt.ylim([min(plt.ylim()), 1])
+        plt.title(f"Training and Validation {key}")
+    plt.show()
+
+
 def main(debug: bool = False, eager: bool = False):
 
     if debug:
@@ -85,24 +103,27 @@ def main(debug: bool = False, eager: bool = False):
         metrics=lambda: elegy.metrics.Accuracy(),
         run_eagerly=eager,
     )
-
+    epochs = 10
     # Fit with datasets in memory
-    model.fit(
+    history = model.fit(
         x=x,
         y=y,
-        epochs=10,
+        epochs=epochs,
         batch_size=64,
+        initial_epoch=epochs * 0,
         steps_per_epoch=100,
         validation_data=(x_val, y_val),
         shuffle=True,
     )
+    plot_history(history)
     # exit()
 
     # Fit with validation from train
-    model.fit(
+    history = model.fit(
         x=x,
         y=y,
-        epochs=10,
+        epochs=2 * epochs,
+        initial_epoch=epochs * 1,
         batch_size=64,
         steps_per_epoch=100,
         validation_split=0.2,
@@ -114,9 +135,10 @@ def main(debug: bool = False, eager: bool = False):
     x = load_dataset("train", is_training=True, batch_size=64, for_fit=True)
     validation = load_dataset("train", is_training=False, batch_size=1000, for_fit=True)
 
-    model.fit(
+    history = model.fit(
         x,
-        epochs=10,
+        epochs=3 * epochs,
+        initial_epoch=epochs * 2,
         steps_per_epoch=100,
         validation_data=validation,
         validation_steps=2,
